+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * [EACQAPW-5208] AppChecker and TinkoffPayAppChecker Tests
 * [EACQAPW-5209] MoneyFormatter Tests
 * [EACQAPW-5238] PaymentController Tests
+* [EACQAPW-5262] CardsController Tests
 
 ### Fixed